Testes No Middleware de Limit Rate
Execução de testes no middleware de limit rate devido a um erro encontrado removi os headers de resposta

mudança para deixar o algoritimo  do jwt explicito

Passei o BD para um container docker, tive que mudar a porta entt adicionei uma variavel ao env referente a porta do bd e passei essa variavel para a instancia sequelize no arquivo /config/BD.js

Algumas mudanças nos Modelos, no modelo de sessoes mudei o tipo do campo Token de String para Text, para ter uma maior segurança em relação ao tamanho dos tokens que serão armazenados, mudança tbm no tipo do campo de preços no modelo Produtos

+++ models/mProdutos.js
@@ -27,7 +27,7 @@ const Produtos = sequelize.define(
         },
 
         Preco : {
-            type : DataTypes.INTEGER,
+            type : DataTypes.DECIMAL,
             allowNull: false
         }
     },
